requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-QRcode==3.2.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pillow==10.4.0
PyMySQL==1.1.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
